Upload picture page half done
